wip: criando tabela de users

+++ db/migrate/20250627031632_create_a_tipo_eventos.rb
@@ -2,17 +2,18 @@
 
 class CreateATipoEventos < ActiveRecord::Migration[7.2]
   def up
-    create_table :a_tipo_eventos do |t|
-      t.string :descricao
-          
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:a_tipo_eventos)
+      create_table :a_tipo_eventos do |t|
+        t.string    :descricao
+        t.string    :created_by
+        t.string    :updated_by
+        t.datetime  :deleted_at
+        t.timestamps
+      end
     end
   end
 
-  def down
-    drop_table :a_tipo_eventos
-  end
+ def down
+  drop_table :a_tipo_eventos if table_exists?(:a_tipo_eventos)
+ end
 end
