+++ db/migrate/20250627031734_create_a_cor.rb
@@ -2,17 +2,18 @@
 
 class CreateACor < ActiveRecord::Migration[7.2]
   def up
-    create_table :a_cor do |t|
-      t.string :descricao
-          
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:a_cor)
+      create_table :a_cor do |t|
+        t.string   :descricao
+        t.string   :created_by
+        t.string   :updated_by
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :a_cor
+    drop_table :a_cor if table_exists?(:a_cor)
   end
 end
