+++ db/migrate/20250627031816_create_a_portes.rb
@@ -2,17 +2,18 @@
 
 class CreateAPortes < ActiveRecord::Migration[7.2]
   def up
-    create_table :a_portes do |t|
-      t.string :descricao
-          
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:a_portes)
+      create_table :a_portes do |t|
+        t.string   :descricao
+        t.string   :created_by
+        t.string   :updated_by
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :a_portes
+    drop_table :a_portes if table_exists?(:a_portes)
   end
 end
