+++ db/migrate/20250627031827_create_a_especies.rb
@@ -2,17 +2,18 @@
 
 class CreateAEspecies < ActiveRecord::Migration[7.2]
   def up
-    create_table :a_especies do |t|
-      t.string :descricao
-          
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:a_especies)
+      create_table :a_especies do |t|
+        t.string   :descricao
+        t.string   :created_by
+        t.string   :updated_by
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :a_especies
+    drop_table :a_especies if table_exists?(:a_especies)
   end
 end
